group all history tab functions in one file
