+++ history_tab.py
@@ -1,25 +1,102 @@
-from clear_audio import clear_audio
 import gradio as gr
 
 
 import glob
 import json
 import shutil
+import os
+
+
+def open_folder(folder_path):
+    os.startfile(folder_path)
+
+def open_outputs_folder():
+    open_folder("outputs")
+
+def open_favorites_folder():
+    open_folder("favorites")
+
+def save_to_favorites(json_text):
+    shutil.copy(json_text["filename"], "favorites/")
+    shutil.copy(json_text["filename_png"], "favorites/")
+    # write new json file with new filename
+    json_text["filename"] = json_text["filename"].replace(
+        "outputs/", "favorites/")
+    json_text["filename_png"] = json_text["filename_png"].replace(
+        "outputs/", "favorites/")
+    json_text["filename_json"] = json_text["filename_json"].replace(
+        "outputs/", "favorites/")
+    with open(json_text["filename_json"].replace("outputs/", "favorites/"), "w") as f:
+        json.dump(json_text, f, indent=2)
+    return gr.Button.update(value="Saved")
+
+
+def delete_generation_cb(refresh):
+    def delete_generation(json):
+        os.remove(json["filename"])
+        os.remove(json["filename_png"])
+        os.remove(json["filename_json"])
+
+        return refresh()
+
+    return delete_generation
+
 
 wav_file_list = []
 wav_file_list_favorites = []
 
-def history_tab(open_outputs_folder, delete_generation_cb):
+
+def get_wav_files():
+    wav_files = glob.glob("outputs/*.wav")
+    global wav_file_list
+    wav_file_list = wav_files
+    return [[file] for file in wav_files]
+
+
+def get_wav_files_favorites():
+    wav_files = glob.glob("favorites/*.wav")
+    global wav_file_list_favorites
+    wav_file_list_favorites = wav_files
+    return [[file] for file in wav_files]
+
+def select_audio(selection):
+    global wav_file_list
+    filename = wav_file_list[selection]
+    with open(filename.replace(".wav", ".json")) as f:
+        json_text = json.load(f)
+    return [
+        gr.Audio.update(value=filename, label=filename),
+        gr.Image.update(value=filename.replace(".wav", ".png")),
+        gr.JSON.update(value=json_text),
+        gr.Button.update(visible=True),
+        gr.Button.update(visible=True, value="Save to favorites"),
+    ]
+
+
+def select_audio_favorites(selection):
+    global wav_file_list_favorites
+    filename = wav_file_list_favorites[selection]
+    with open(filename.replace(".wav", ".json")) as f:
+        json_text = json.load(f)
+    return [
+        gr.Audio.update(value=filename, label=filename),
+        gr.Image.update(value=filename.replace(".wav", ".png")),
+        gr.JSON.update(value=json_text),
+        gr.Button.update(visible=True),
+    ]
+
+def clear_audio(): return [
+    gr.Audio.update(value=None, label=""),
+    gr.Image.update(value=None),
+    gr.JSON.update(value=None),
+    gr.Button.update(visible=False),
+]
+
+def history_tab():
     with gr.Tab("History") as history_tab:
         with gr.Row():
             button_output = gr.Button(value="Open outputs folder")
         button_output.click(open_outputs_folder)
-
-        def get_wav_files():
-            wav_files = glob.glob("outputs/*.wav")
-            global wav_file_list
-            wav_file_list = wav_files
-            return [[file] for file in wav_files]
 
         _audio_component = gr.Audio(visible=False)
         history_list = gr.Dataset(components=[_audio_component],
@@ -29,20 +106,6 @@ def history_tab(open_outputs_folder, delete_generation_cb):
         history_image = gr.Image()
         history_json = gr.JSON()
 
-        def save_to_favorites(json_text):
-            shutil.copy(json_text["filename"], "favorites/")
-            shutil.copy(json_text["filename_png"], "favorites/")
-            # write new json file with new filename
-            json_text["filename"] = json_text["filename"].replace(
-                "outputs/", "favorites/")
-            json_text["filename_png"] = json_text["filename_png"].replace(
-                "outputs/", "favorites/")
-            json_text["filename_json"] = json_text["filename_json"].replace(
-                "outputs/", "favorites/")
-            with open(json_text["filename_json"].replace("outputs/", "favorites/"), "w") as f:
-                json.dump(json_text, f, indent=2)
-            return gr.Button.update(value="Saved")
-
         with gr.Row():
             delete_from_history = gr.Button(
                 value="Delete", variant="stop", visible=False)
@@ -51,19 +114,6 @@ def history_tab(open_outputs_folder, delete_generation_cb):
 
         save_to_favorites_history.click(
             fn=save_to_favorites, inputs=history_json, outputs=save_to_favorites_history)
-
-        def select_audio(selection):
-            global wav_file_list
-            filename = wav_file_list[selection]
-            with open(filename.replace(".wav", ".json")) as f:
-                json_text = json.load(f)
-            return [
-                gr.Audio.update(value=filename, label=filename),
-                gr.Image.update(value=filename.replace(".wav", ".png")),
-                gr.JSON.update(value=json_text),
-                gr.Button.update(visible=True),
-                gr.Button.update(visible=True, value="Save to favorites"),
-            ]
 
         history_list.select(fn=select_audio, inputs=[history_list], outputs=[
                             history_audio, history_image, history_json, delete_from_history, save_to_favorites_history], preprocess=False)
@@ -78,17 +128,11 @@ def history_tab(open_outputs_folder, delete_generation_cb):
         history_tab.select(fn=update_history_tab, outputs=[history_list])
 
 
-def favorites_tab(open_favorites_folder, delete_generation_cb):
+def favorites_tab():
     with gr.Tab("Favorites") as favorites_tab:
         with gr.Row():
             button_output = gr.Button(value="Open favorites folder")
         button_output.click(open_favorites_folder)
-
-        def get_wav_files_favorites():
-            wav_files = glob.glob("favorites/*.wav")
-            global wav_file_list_favorites
-            wav_file_list_favorites = wav_files
-            return [[file] for file in wav_files]
 
         _audio_component = gr.Audio(visible=False)
         favorites_list = gr.Dataset(components=[_audio_component],
@@ -102,19 +146,7 @@ def favorites_tab(open_favorites_folder, delete_generation_cb):
             delete_from_favorites = gr.Button(
                 value="Delete", variant="stop", visible=False)
 
-        def select_audio(selection):
-            global wav_file_list_favorites
-            filename = wav_file_list_favorites[selection]
-            with open(filename.replace(".wav", ".json")) as f:
-                json_text = json.load(f)
-            return [
-                gr.Audio.update(value=filename, label=filename),
-                gr.Image.update(value=filename.replace(".wav", ".png")),
-                gr.JSON.update(value=json_text),
-                gr.Button.update(visible=True),
-            ]
-
-        favorites_list.select(fn=select_audio, inputs=[favorites_list], outputs=[
+        favorites_list.select(fn=select_audio_favorites, inputs=[favorites_list], outputs=[
             favorites_audio, favorites_image, favorites_json, delete_from_favorites], preprocess=False)
 
         def update_favorites_tab():
