+++ server.py
@@ -6,7 +6,6 @@ import matplotlib.pyplot as plt
 from generation_tab_bark import generation_tab_tortoise, generation_tab_bark
 import gradio as gr
 import json
-import os
 from history_tab import favorites_tab, history_tab
 from model_manager import model_manager
 from settings_tab import settings_tab
@@ -63,32 +62,14 @@ def load_models(
     return gr.Button.update(value="Reload models", interactive=True)
 
 
-def open_outputs_folder():
-    os.startfile("outputs")
-
-
-def open_favorites_folder():
-    os.startfile("favorites")
-
-
-def delete_generation_cb(refresh):
-    def delete_generation(json):
-        os.remove(json["filename"])
-        os.remove(json["filename_png"])
-        os.remove(json["filename_json"])
-
-        return refresh()
-
-    return delete_generation
 
 with gr.Blocks() as block:
     gr.Markdown("# TTS Generation WebUI (Bark & Tortoise)")
     generation_tab_bark()
     generation_tab_tortoise()
 
-    history_tab(open_outputs_folder, delete_generation_cb)
-
-    favorites_tab(open_favorites_folder, delete_generation_cb)
+    history_tab()
+    favorites_tab()
 
     settings_tab(config, save_config, load_models)
